unit tests working
